+++ app/src/main/java/org/simple/clinic/home/HomeScreen.kt
@@ -13,7 +13,9 @@ import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.tabs.TabLayoutMediator
 import com.jakewharton.rxbinding3.view.clicks
 import com.spotify.mobius.functions.Consumer
+import io.reactivex.Observable
 import io.reactivex.rxkotlin.cast
+import io.reactivex.subjects.PublishSubject
 import kotlinx.parcelize.Parcelize
 import org.simple.clinic.R
 import org.simple.clinic.ReportAnalyticsEvents
@@ -75,9 +77,6 @@ class HomeScreen :
   @Inject
   lateinit var features: Features
 
-  @Inject
-  lateinit var homeScreenUpdate: HomeScreenUpdate
-
   private val homeScreenRootLayout
     get() = binding.homeScreenRootLayout
 
@@ -96,6 +95,8 @@ class HomeScreen :
   private val facilitySelectButton
     get() = binding.facilitySelectButton
 
+  private val hotEvents = PublishSubject.create<HomeScreenEvent>()
+
   private val tabs = listOf(PATIENTS, OVERDUE, REPORTS)
 
   override fun defaultModel() = HomeScreenModel.create()
@@ -105,11 +106,15 @@ class HomeScreen :
   override fun bindView(layoutInflater: LayoutInflater, container: ViewGroup?) =
       ScreenHomeBinding.inflate(layoutInflater, container, false)
 
-  override fun events() = facilitySelectionClicks()
+  override fun events() = Observable
+      .mergeArray(
+          facilitySelectionClicks(),
+          hotEvents
+      )
       .compose(ReportAnalyticsEvents())
       .cast<HomeScreenEvent>()
 
-  override fun createUpdate() = homeScreenUpdate
+  override fun createUpdate() = HomeScreenUpdate()
 
   override fun createInit() = HomeScreenInit()
 
@@ -151,6 +156,10 @@ class HomeScreen :
   override fun onDestroyView() {
     viewPager.adapter = null
     super.onDestroyView()
+  }
+
+  fun overdueListCountUpdated(count: Int) {
+    hotEvents.onNext(OverdueAppointmentCountUpdated(count))
   }
 
   private fun setupToolBar() {